Prepare code for separation of parser from core
Add lexical stack frames so that the code can have a better
understanding of what is being parsed. Not quite a full AST
but the equivalent thereof.

We can now begin to separate the processes of reading the input and
writing the output. This is crucial if we want to introduce processors
that do not write HTML.